flytt behandling av pdl-data til service-klassen. Bruk egen graphql-query for å hente navn
