+++ src/main/kotlin/no/nav/familie/ba/soknad/api/personopplysning/PdlClient.kt
@@ -1,13 +1,17 @@
 package no.nav.familie.ba.soknad.api.personopplysning
 
+import com.fasterxml.jackson.databind.JsonNode
 import no.nav.familie.http.client.AbstractRestClient
 import no.nav.familie.http.client.Pingable
 import org.apache.commons.lang3.StringUtils
+import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.beans.factory.annotation.Value
+import org.springframework.http.HttpMethod
 import org.springframework.stereotype.Component
 import org.springframework.web.client.RestOperations
+import org.springframework.web.client.exchange
 import java.net.URI
 
 
@@ -16,11 +20,10 @@ class PdlClient(@Value("\${PDL_API_URL}") private val pdlBaseUrl: String,
                 @Qualifier("restKlientMedApiKey") private val restOperations: RestOperations)
     : AbstractRestClient(restOperations, "integrasjon"), Pingable {
 
-    val pdlUri: URI = URI.create("$pdlBaseUrl/graphql")
-    private val personInfoQuery = this::class.java.getResource("/pdl/hentperson-med-relasjoner.graphql").readText().graphqlCompatible()
+    private val pdlUri: URI = URI.create("$pdlBaseUrl/graphql")
 
-    private fun hentPersonData(personIdent: String): PdlHentPersonResponse {
-        val pdlPersonRequest = PdlPersonRequest(variables = PdlPersonRequestVariables(personIdent), query = personInfoQuery)
+    private fun hentPersonData(personIdent: String, query: String): PdlHentPersonResponse {
+        val pdlPersonRequest = PdlPersonRequest(variables = PdlPersonRequestVariables(personIdent), query = query)
         try {
             val response = postForEntity<PdlHentPersonResponse>(uri = pdlUri, payload = pdlPersonRequest)
             if (!response.harFeil()) {
@@ -34,47 +37,28 @@ class PdlClient(@Value("\${PDL_API_URL}") private val pdlBaseUrl: String,
         }
     }
 
-    private fun hentNavn(personIdent: String): String {
-        val response = hentPersonData(personIdent)
-        return Result.runCatching {
-            response.data.person!!.navn.first().fulltNavn()
-        }.fold(
-                onSuccess = { it },
-                onFailure = { throw it}
-        )
+    fun hentNavn(personIdent: String): PdlHentPersonResponse {
+        val query = this::class.java.getResource("/pdl/hentnavn.graphql").readText().graphqlCompatible()
+        return hentPersonData(personIdent, query)
     }
 
-    fun hentPerson(personIdent: String): Person {
-        val response = hentPersonData(personIdent)
-        return Result.runCatching {
-            val barn: Set<Barn> = response.data.person!!.familierelasjoner.filter {
-                relasjon -> relasjon.relatertPersonsRolle == FAMILIERELASJONSROLLE.BARN
-            }.map {
-                relasjon -> Barn(relasjon.relatertPersonsIdent, hentNavn(relasjon.relatertPersonsIdent))
-            }.toSet()
-
-            response.data.person.let {
-                Person(navn = it.navn.first().fulltNavn(), barn = barn)
-            }
-        }.fold(
-                onSuccess = { it },
-                onFailure = { throw it}
-        )
+    fun hentNavnOgRelasjoner(personIdent: String): PdlHentPersonResponse {
+        val query = this::class.java.getResource("/pdl/hentperson-med-relasjoner.graphql").readText().graphqlCompatible()
+        return hentPersonData(personIdent, query)
     }
 
-    // TODO: fjerne utkommentering
     override fun ping() {
-        /*try {
+        try {
             restOperations.exchange<JsonNode>(pdlUri, HttpMethod.OPTIONS)
             LOG.debug("Ping mot PDL-API OK")
         } catch (e: Exception) {
             LOG.warn("Ping mot PDL-API feilet")
             throw IllegalStateException("Ping mot PDL-API feilet", e)
-        }*/
+        }
     }
 
     companion object {
-        private val LOG = LoggerFactory.getLogger(PdlClient::class.java)
+        val LOG: Logger = LoggerFactory.getLogger(PdlClient::class.java)
     }
 }
