+++ src/main/kotlin/no/nav/familie/ba/soknad/api/personopplysning/Person.kt
@@ -9,6 +9,3 @@ data class Barn(
         val ident: String,
         val navn: String
 )
-
-
-
